requirements.txt
# Core Django
Django==5.2.7
python-dotenv==1.0.0

# Database
PyMySQL==1.1.2

# Django Apps
django-ckeditor==6.7.3
django-recaptcha==4.1.0

# Image handling
Pillow==12.0.0

# Firebase
firebase-admin==6.5.0

# Email Marketing
mailchimp-marketing==3.0.80

# HTTP Requests
requests==2.32.5

# SMS (if you're using Twilio)
twilio==9.8.7

# Phone number validation (if used)
phonenumbers==9.0.19

# Development only (remove for production or use separate requirements-dev.txt)
django-debug-toolbar==6.1.0

#django OAUTH
django-allauth==65.13.1

#django Paypal
django-paypal==2.1